Add files via upload
create two pages

They will both have a centered container id'd "main", at 800 pixels wide with 20 pixels of padding, and a 1 pixel black border.

In the container on each page, add three paragraphs of lorem ipsum text. 

At the bottom of each page, create a link to page 2 on page 1 and a link to page 1 on page 2. 

index.html - has link that says "Next page" that directs the user to page 2.
page2.html - has link that says "Previous page" that directs the user to page 1(index).